+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - OAuth merged
 - Fix Grafana Cloud login
 - Updated modules
-- Add "window-position" option
+- Added "window-position" option, allows running kiosk on different displays
+- Added `--check-for-update-interval=31536000` to default flags sent to chromium to workaround update popup
 
 ## 1.0.2